+++ deployment/CHANGELOG.md
@@ -7,6 +7,8 @@
 
 ### Bugfixes
 
+- [#116](https://github.com/fragmuffin/cqparts/pull/116): bugfix for import unicode filenames
+
 
 ----
 
@@ -17,3 +19,5 @@
 - [#83](https://github.com/fragmuffin/cqparts/pull/83): [`Stepper`](https://fragmuffin.github.io/cqparts/doc/api/cqparts_motors#cqparts_motors.stepper.Stepper) motor (author: [@zignig](https://github.com/zignig))
 
 ### Bugfixes
+
+- [#114](https://github.com/fragmuffin/cqparts/pull/114): added cutout and fixed stepper construct (author: [@zignig](https://github.com/zignig))